move document routes from main to dedicated file
