+++ backend/api/__init__.py
@@ -78,10 +78,11 @@ def create_app(test_config=None):
     Migrate(app, db)
 
     # import and register blueprints
-    from api.views import main, message, box, fp, pm, docclass
+    from api.views import main, document, message, box, fp, pm, docclass
 
     # why blueprints http://flask.pocoo.org/docs/1.0/blueprints/
     app.register_blueprint(main.main)
+    app.register_blueprint(document.document)
     app.register_blueprint(message.message)
     app.register_blueprint(box.box)
     app.register_blueprint(fp.fp)
